app/requirements.txt
torch
torchvision
pillow
fastapi
uvicorn
mysql-connector-python
requests
python-multipart